abstract-algebra: add common examples
For groups, rings and fields, add common examples like
integers and real numbers.

This is necessary because of the following situation:
Theorem A proves a property of fields.
Theorem B requires theorem A, but only for the field of real numbers.
So we need to mention somewhere that real numbers form a field.